feat: 拆分类型定义文件
